backend/pyproject.toml
[project]
name = "image-solve-backend"
version = "1.0.0"
description = "Image Solve Platform MVP Backend"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14,<4.0",
    "psycopg2-binary>=2.9,<3.0",
    "argon2-cffi>=23.0,<24.0",
    "drf-spectacular>=0.27,<0.28",
    "django-cors-headers>=4.3,<5.0",
    "gunicorn>=21.0,<22.0",
    "requests>=2.31,<3.0",
    "dj-database-url>=3.0.1",
    "Pillow>=10.0,<11.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "django-stubs[compatible-mypy]>=4.2",
    "djangorestframework-stubs>=3.14",
    "types-requests>=2.31",
    "ruff>=0.1.9",
    "black>=23.12",
    "pytest>=7.4",
    "pytest-django>=4.7",
    "model-bakery>=1.17",
    "requests-mock>=1.11",
]
redis = [
    "redis>=5.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["config", "customers", "usage", "core"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["*/migrations/*.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.2.2",
    "djangorestframework-stubs>=3.16.1",
    "model-bakery>=1.20.5",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "requests-mock>=1.12.1",
    "ruff>=0.12.8",
    "types-requests>=2.32.4.20250809",
]
